refactor!: rename offline_analyzers/ → analyzers/
Historical name dating from when an ``online_analyzers/`` counterpart
was contemplated. No online analyzers ever materialised; the
``offline_`` qualifier has been redundant from day one.

This commit:
- ``git mv image_analysis/offline_analyzers image_analysis/analyzers``
- Sweep-replaces ``image_analysis.offline_analyzers`` →
  ``image_analysis.analyzers`` across all import sites (33 files in
  GEECS-Plugins).

Class paths in YAMLs (in GEECS-Plugins-Configs) follow in the
companion configs PR.
