+++ ImageAnalysis/image_analysis/analyzers/Undulator/BCaveMagSpecStitcher.py
@@ -16,7 +16,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 from image_analysis.tools.rendering import base_render_image
-from image_analysis.offline_analyzers.standard_analyzer import StandardAnalyzer
+from image_analysis.analyzers.standard_analyzer import StandardAnalyzer
 from image_analysis.config.array2d_processing import CameraConfig
 from image_analysis.types import AnalyzerResultDict
