+++ ImageAnalysis/image_analysis/analyzers/Undulator/hi_res_mag_cam_analyzer.py
@@ -15,7 +15,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 from image_analysis.tools.rendering import base_render_image
-from image_analysis.offline_analyzers.beam_analyzer import BeamAnalyzer
+from image_analysis.analyzers.beam_analyzer import BeamAnalyzer
 from image_analysis.config.array2d_processing import CameraConfig
 from image_analysis.algorithms.bowtie_fit import BowtieFitAlgorithm
 from image_analysis.types import ImageAnalyzerResult
