+++ ImageAnalysis/image_analysis/analyzers/grenouille_analyzer.py
@@ -21,7 +21,7 @@ import numpy as np
 import pandas as pd
 
 # Import the StandardAnalyzer parent class
-from image_analysis.offline_analyzers.standard_analyzer import StandardAnalyzer
+from image_analysis.analyzers.standard_analyzer import StandardAnalyzer
 from image_analysis.config.array2d_processing import CameraConfig
 
 from image_analysis.algorithms.frog_dll_retrieval import (
