+++ ImageAnalysis/image_analysis/analyzers/line_analyzer.py
@@ -18,7 +18,7 @@ from typing import Optional, Dict
 
 
 # Import the Standard1DAnalyzer parent class
-from image_analysis.offline_analyzers.standard_1d_analyzer import Standard1DAnalyzer
+from image_analysis.analyzers.standard_1d_analyzer import Standard1DAnalyzer
 from image_analysis.config.array1d_processing import Line1DConfig
 
 # Import line-specific tools
