+++ ImageAnalysis/image_analysis/analyzers/line_stitcher.py
@@ -20,7 +20,7 @@ from typing import Dict, List, Optional
 import numpy as np
 
 from image_analysis.data_1d_utils import read_1d_data
-from image_analysis.offline_analyzers.line_analyzer import LineAnalyzer
+from image_analysis.analyzers.line_analyzer import LineAnalyzer
 from image_analysis.config.array1d_processing import Line1DConfig
 from image_analysis.types import Array1D, ImageAnalyzerResult
