+++ tests/integration/data/scan_analysis/test_array1d_scan_analyzer.py
@@ -27,7 +27,7 @@ def test_array1d_ict_noscan_runs():
     """
     from geecs_data_utils import ScanTag
     from image_analysis.config.loader import load_line_config
-    from image_analysis.offline_analyzers.standard_1d_analyzer import Standard1DAnalyzer
+    from image_analysis.analyzers.standard_1d_analyzer import Standard1DAnalyzer
     from scan_analysis.analyzers.common.array1d_scan_analysis import Array1DScanAnalyzer
 
     image_analyzer = Standard1DAnalyzer(line_config=load_line_config(ICT_DEV))
@@ -48,7 +48,7 @@ def test_array1d_ict_results_are_2d():
     """Each per-shot ICT result contains a valid Nx2 line_data array."""
     from geecs_data_utils import ScanTag
     from image_analysis.config.loader import load_line_config
-    from image_analysis.offline_analyzers.standard_1d_analyzer import Standard1DAnalyzer
+    from image_analysis.analyzers.standard_1d_analyzer import Standard1DAnalyzer
     from scan_analysis.analyzers.common.array1d_scan_analysis import Array1DScanAnalyzer
 
     image_analyzer = Standard1DAnalyzer(line_config=load_line_config(ICT_DEV))
@@ -80,7 +80,7 @@ def test_array1d_magspec_noscan_runs():
     """
     from geecs_data_utils import ScanTag
     from image_analysis.config.loader import load_line_config
-    from image_analysis.offline_analyzers.line_analyzer import LineAnalyzer
+    from image_analysis.analyzers.line_analyzer import LineAnalyzer
     from scan_analysis.analyzers.common.array1d_scan_analysis import Array1DScanAnalyzer
 
     image_analyzer = LineAnalyzer(line_config=load_line_config(MAGSPEC_DEV))
@@ -101,7 +101,7 @@ def test_array1d_magspec_results_have_com_scalar():
     """Each per-shot MagSpec result contains the CoM scalar from LineAnalyzer."""
     from geecs_data_utils import ScanTag
     from image_analysis.config.loader import load_line_config
-    from image_analysis.offline_analyzers.line_analyzer import LineAnalyzer
+    from image_analysis.analyzers.line_analyzer import LineAnalyzer
     from scan_analysis.analyzers.common.array1d_scan_analysis import Array1DScanAnalyzer
 
     image_analyzer = LineAnalyzer(line_config=load_line_config(MAGSPEC_DEV))
